fix: 背面剔除Y/Z面修正+ECS 16-bit+流体渲染+SKIP_BLOCKS
- 背⾯剔除 Y/Z 轴面分配校正，阈值>30K方块，仅整体视图生效
- ECS 16-bit 调色板——palSize>256自动切换16-bit索引，8-bit旧格式兼容
- 流体源方块半透明渲染（水蓝岩浆橙色），tesselator+position_color shader
- 扫描跳过20种无法独立存在方块，公开skipBlock() API
- 流动液体视为空气，shade亮度0.6→1.0，重复绘制清理

+++ src/main/java/com/endlessepoch/ecsformat/EcsRawData.java
@@ -7,11 +7,11 @@ public class EcsRawData {
     public final int width, height, depth;
     public final int controllerX, controllerY, controllerZ;
     public final List<EcsPaletteEntry> palette;
-    public final byte[] voxelData;
+    public final short[] voxelData;
 
     public EcsRawData(int width, int height, int depth,
                       int controllerX, int controllerY, int controllerZ,
-                      List<EcsPaletteEntry> palette, byte[] voxelData) {
+                      List<EcsPaletteEntry> palette, short[] voxelData) {
         this.width = width;
         this.height = height;
         this.depth = depth;
